requirements.txt
nibabel==5.2.1
numpy==1.24.1
opencv-python==4.10.0.84
pandas==2.2.2
Pillow==9.0.1
scipy==1.14.1
torch==2.0.1
torchvision==0.15.2
